Fix regression in #include directive paths
These `#include` directives were modified to expect the target files in the same folder as the containing file, but
those files are not in that folder. This caused compilation of the library to fail:

c:\Users\per\Documents\Arduino\libraries\BSEC_Software_Library\src/bsec.h:46:10: fatal error: bsec_datatypes.h: No such file or directory
^~~~~~~~~~~~~~~~~~

The target files are in subfolders relative to the file containing the `#include` directive, so that relative path must
be specified in the `#include` directive as before.